Implement hittable surfaces

+++ src/vec3.rs
@@ -2,28 +2,28 @@ use std::fmt::Display;
 use std::ops::{Add, Div, Mul, Sub};
 #[derive(Copy, Clone)]
 pub struct Vec3 {
-    x: f32,
-    y: f32,
-    z: f32,
+    x: f64,
+    y: f64,
+    z: f64,
 }
 
 impl Vec3 {
-    pub fn new(x: f32, y: f32, z: f32) -> Self {
+    pub fn new(x: f64, y: f64, z: f64) -> Self {
         Vec3 { x, y, z }
     }
 
-    pub fn x(self) -> f32 {
+    pub fn x(self) -> f64 {
         self.x
     }
 
-    pub fn y(self) -> f32 {
+    pub fn y(self) -> f64 {
         self.y
     }
-    pub fn z(self) -> f32 {
+    pub fn z(self) -> f64 {
         self.z
     }
 
-    pub fn dot(&self, other: &Vec3) -> f32 {
+    pub fn dot(&self, other: &Vec3) -> f64 {
         self.x * other.x + self.y * other.y + self.z * other.z
     }
 
@@ -35,12 +35,12 @@ impl Vec3 {
         }
     }
 
-    pub fn len(&self) -> f32 {
-        f32::sqrt(self.len_squared())
+    pub fn len(&self) -> f64 {
+        f64::sqrt(self.len_squared())
     }
 
-    pub fn len_squared(&self) -> f32 {
-        f32::powf(self.x, 2.0) + f32::powf(self.y, 2.0) + f32::powf(self.z, 2.0)
+    pub fn len_squared(&self) -> f64 {
+        f64::powf(self.x, 2.0) + f64::powf(self.y, 2.0) + f64::powf(self.z, 2.0)
     }
 
     pub fn unit_vector(&self) -> Vec3 {
@@ -79,10 +79,10 @@ impl Sub for Vec3 {
         }
     }
 }
-impl Mul<f32> for &Vec3 {
+impl Mul<f64> for &Vec3 {
     type Output = Vec3;
 
-    fn mul(self, rhs: f32) -> Self::Output {
+    fn mul(self, rhs: f64) -> Self::Output {
         Vec3 {
             x: self.x * rhs,
             y: self.y * rhs,
@@ -90,10 +90,10 @@ impl Mul<f32> for &Vec3 {
         }
     }
 }
-impl Mul<f32> for Vec3 {
+impl Mul<f64> for Vec3 {
     type Output = Self;
 
-    fn mul(self, rhs: f32) -> Self::Output {
+    fn mul(self, rhs: f64) -> Self::Output {
         Vec3 {
             x: self.x * rhs,
             y: self.y * rhs,
@@ -101,7 +101,7 @@ impl Mul<f32> for Vec3 {
         }
     }
 }
-impl Mul<&Vec3> for f32 {
+impl Mul<&Vec3> for f64 {
     type Output = Vec3;
 
     fn mul(self, rhs: &Vec3) -> Self::Output {
@@ -113,7 +113,7 @@ impl Mul<&Vec3> for f32 {
     }
 }
 
-impl Mul<Vec3> for f32 {
+impl Mul<Vec3> for f64 {
     type Output = Vec3;
 
     fn mul(self, rhs: Vec3) -> Self::Output {
@@ -124,10 +124,10 @@ impl Mul<Vec3> for f32 {
         }
     }
 }
-impl Div<f32> for Vec3 {
+impl Div<f64> for Vec3 {
     type Output = Self;
 
-    fn div(self, rhs: f32) -> Self::Output {
+    fn div(self, rhs: f64) -> Self::Output {
         self * (1.00 / rhs)
     }
 }
